requirements.txt

pandas
openpyxl
openai
tqdm
PyYAML
